refactor: move `sway::Client` out of compositor
The module `compositor` is gated behind the `workspaces` feature, but
`sway::Client` is no longer used only by the `workspaces` module.
